Add support for highlighted material.
Instead of puzzling out the #pdfloc data, we use the marked_text to find
the text that should be highlighted.  This will have some problems:
 - If the highlighted text appears multiple times on a page, we won't
   know which one to highlight.
 - We have to produce text in the same way as the reader does, and we
   don't know how that works.  So far, we know:
    * Spaces disappear
    * New lines become spaces
    * Hyphens + new lines disappear
    * Ligatures are turned back into normal letters
   But I'm sure there are more complications lying in wait.
We now use PDFMiner to get the text (with positions) off of the page, and
the custom highlighter from the highlight branch to draw the highlighted
regions on the page.

Highlights with text notes work; highlights with drawings do not.  This
mostly solves bug #1.