Work out cropBox vs. mediaBox issues
PDF viewers seem to show only the intersection of these two boxes.  So
everywhere we need to figure out the visible area, we'll use the the
intersection.  Currently, we behave badly if the intersection is empty,
but viewers don't like that either, so this shouldn't be too much of a
problem.

Also, note that PDFMiner reports locations relative to the mediaBox,
regardless of how it interacts with the cropBox.