manager.py: Allow syncing without an original file
When a file is imported into the library, there may not be an original
file on the computer.  This is not a problem for syncing, as there is no
dice map in this case, so the PDF file on the reader is just as good of
a source.  So if there is no original file but also no dice map, we
should just use the reader file as the source for constructing the
annotated PDF.