Switch to custom SVG to PDF renderer
We only need to support a small fraction of the SVG spec, namely
solidly-stroked polylines.  By doing this ourselves, we can eliminate
carrying a forked version of svglib and remove the dependency on
reportlab.

With the pdf_add_content method, we can add arbitrary content to a PDF
page, which may prove to be useful in the future.