pdfannotation: Add explicit popups for text annotation
Evince doesn't create popup boxes for text annotations unless they are
explicitly included in the annotation object.  So we do this.  We need
to refer to the popup (and from the popup back to the text annotation)
with indirect objects, so we put off dealing with the popup annotations
until the add_annotation method, since we need the PDF object to make
the indirect objects.

As a gotcha, acroread won't display the popups unless they are
explicitly added to the annotations list.  (Evince isn't so picky.)