Ignore spaces in when matching highlighted text
The database removes spaces from highlighted text, but sometimes puts
spaces at newlines.  We had been attempting to replicate that when
matching the text, but it seems to only work sometimes.  Now we remove
spaces from both the database text and the page text before trying to
match the two.