Compatibility with the new PRS-T2
Only two changes are needed: The default path on the reader for storing
files is now Sony_Reader/media/books (instead of downloads), since that
exists on both the T1 and the T2.  A slight change to an SQL statement
accounts for the fact that the markups table now has a mime_type column.

Thanks to maptau_07 of the MobileRead forums for troubleshooting.